requirements.txt
lightning==2.0.0
pandas==1.5.3
google-cloud-storage==2.8.0
pytest==7.2.2
jupyterlab==3.6.2
